+++ packages/watcher/app.py
@@ -1,0 +1,18 @@
+from flask import Flask, request
+from config import DEV_CHAT_ID
+from bot import send_message, watcherbot
+from logger import logger
+
+
+app = Flask(__name__)
+
+@app.route('/failure-report', methods=['POST'])
+def backend_failure() -> tuple:
+    system_name = request.json.get('system_name', '')
+    error_message = request.json.get('error_message', '')
+    
+    send_message(DEV_CHAT_ID, f'{system_name} send an error message: {error_message}')
+    logger.error(f'Received {system_name} failure: {error_message}')
+    
+    return {'status': 'success'}, 200
+
